Notificacao[OK] e Agendamentos [2]

+++ UBeauty/src/main/java/com/ubeauty/Entities/Cupom.java
@@ -58,7 +58,7 @@ public class Cupom implements Serializable {
         this.nome = nome;
     }
 
-    public double getDesconto() {
+    public int getDesconto() {
         return desconto;
     }
